+++ package.json
@@ -3,14 +3,20 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^4.2.4",
+    "@testing-library/jest-dom": "^5.15.0",
     "@testing-library/react": "^9.3.2",
     "@testing-library/user-event": "^7.1.2",
+    "@types/react-router": "^5.1.17",
+    "axios": "^0.24.0",
     "classnames": "^2.3.1",
     "jsdoc": "^3.6.6",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
-    "react-scripts": "3.4.3"
+    "react-router": "^5.2.1",
+    "react-router-dom": "^6.0.1",
+    "react-scripts": "^4.0.3",
+    "socket.io": "^4.3.1",
+    "socket.io-client": "^4.3.2"
   },
   "scripts": {
     "start": "react-scripts start",
